# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Auth service",
        "title": "AuthService API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/auth/change-password": {
            "post": {
                "description": "Ushbu endpoint foydalanuvchining parolini o'zgartiradi.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Parolni o'zgartirish",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Access token",
                        "name": "Acces-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Parolni o'zgartirish uchun foydalanuvchi ma'lumotlari",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ChangePassReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Parol muvaffaqiyatli o'zgartirildi",
                        "schema": {
                            "$ref": "#/definitions/model.ChangePassResp"
                        }
                    },
                    "400": {
                        "description": "Xatolik yuz berdi",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "description": "Ushbu endpoint foydalanuvchining email va parolini tekshiradi va access hamda refresh tokenlarini qaytaradi.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Foydalanuvchi tizimga kirishi",
                "parameters": [
                    {
                        "description": "Kirish ma'lumotlari (email va parol)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.LoginReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Access va Refresh tokenlar qaytarildi",
                        "schema": {
                            "$ref": "#/definitions/model.LoginResp"
                        }
                    },
                    "400": {
                        "description": "Xatolik yuz berdi",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    },
                    "500": {
                        "description": "Server xatosi",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "description": "Ushbu endpoint foydalanuvchini tizimdan chiqaradi va access tokenni blacklisting ro'yxatiga qo'shadi.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Tizimdan chiqish",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Access token",
                        "name": "Acces-Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tizimdan muvaffaqiyatli chiqdingiz, biz sizni yana kutamiz.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Xatolik yuz berdi",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "Ushbu endpoint foydalanuvchini ro'yxatdan o'tkazadi. Email va parol bilan ma'lumotlarni yuboring.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Foydalanuvchini ro'yxatdan o'tkazish",
                "parameters": [
                    {
                        "description": "Ro'yxatdan o'tish ma'lumotlari",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.RegisterReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Foydalanuvchi muvaffaqiyatli ro'yxatdan o'tkazildi",
                        "schema": {
                            "$ref": "#/definitions/model.RegisterResp"
                        }
                    },
                    "400": {
                        "description": "Xatolik yuz berdi",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    },
                    "500": {
                        "description": "Server xatosi",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
        "/auth/reset-password": {
            "post": {
                "description": "Ushbu endpoint foydalanuvchining parolini qayta tiklaydi.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Parolni qayta tiklash",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Access token",
                        "name": "Acces-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Yangi parol",
                        "name": "password",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Parol muvaffaqiyatli tiklandi",
                        "schema": {
                            "$ref": "#/definitions/model.ResetPassResp"
                        }
                    },
                    "400": {
                        "description": "Xatolik yuz berdi",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.ChangePassReq": {
            "type": "object",
            "properties": {
                "new_password": {
                    "type": "string"
                },
                "now_password": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "model.ChangePassResp": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "model.Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "model.LoginReq": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "model.LoginResp": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "model.RegisterReq": {
            "type": "object",
            "properties": {
                "emil": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "model.RegisterResp": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "model.ResetPassResp": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}